+++ src/SmartArray.php
@@ -234,12 +234,6 @@ class SmartArray extends SmartArrayBase
     }
 
     /** {@inheritDoc} */
-    public function sprintf(string $format): SmartArray
-    {
-        return parent::sprintf($format);
-    }
-
-    /** {@inheritDoc} */
     public function map(callable $callback): static
     {
         return parent::map($callback);